backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-jose
bcrypt
passlib
sentence-transformers
faiss-cpu
pydantic[email]
httpx
pytest
pre-commit
flake8
black